[SYSTEMML-2016] Performance multi-threaded matrix multiply thresholds
Recently, we've introduced a common thread pool for all multi-threaded
operations, which eliminated the costs for thread pool creation per
operation. Therefore, multi-threading is now beneficial for smaller data
sizes. This patch accordingly updates the size and FLOP thresholds used
for matrix multiply (mm) and mmchain operations. In addition, we now
also use a slightly more efficient approach for aggregating partial
results for mmchain.

For example, on a scenario of 10,000 operations, this patch improved the
cumulative runtime for [1Kx1K] mm [1Kx1] from 8.1s to 1.7s, and for
[500x1K] mmchain [1Kx1] from 6.5s to 4s.