[SYSTEMML-2217] Fix OOM read sparse textcell matrix w/ unknown nnz
In case of forced single-node execution, the read of ultra-sparse
textcell matrices with unknown nnzs can cause unnecessary out-of-memory
errors because we always allocate the output in dense format for
unknowns. This patch fixes this robustness issue by obtaining a rough
estimate of the number of non-zeros based on the ratio of filesize on
hdfs to expected size in dense format.