[SYSTEMML-2220] Performance ultra-sparse block matrix mult operations
This patch improves the performance of ultra-sparse matrix multiply
operations with ultra-sparse lhs matrices, which have rows that only
contain 1 non-zero values. In this special case we simply copy the
related rhs row into the sparse output (with optional scaling) because
this structure guarantees that there is no need for aggregation.