[SYSTEMML-2229] Add missing support for rowProds/colProds aggregates
So far we only support full aggregates for products but not the
row/column-wise aggregates as all other unary aggregates. This patch
adds the missing compiler and runtime support for all backends. Similar
as before these operations are not NaN aware by using early out as soon
the temporary output contains a zero but the runtime is configurable in
that regard.

Furthermore, this also includes an improvement of the dense block
abstraction to make the API more fluent for initialization.