[SYSTEMML-2234] Fix buffer pool memory/file leak (missing cleanup)
This patch fixes severe issues of missing variable cleanup for special
cases of function calls and update-in-place variables. The missing
cleanup leads to unnecessary memory pressure in the buffer pool and thus
unnecessary evictions as well as missing cleanup of local and hdfs files
which can lead to "no space left on device" issues on long running
scripts with many variables (e.g., DL with many epochs).

Furthermore, this patch also includes a hardening of the bufferpool
threshold management and thread-safe buffer pool logging.