[SYSTEMML-2238] Reuse of local parfor fair scheduler pool names
This patch fixes a leak of fair scheduler pools in scripts with
repeatedly executed or simply many parfor loops by reusing scheduler
pool names. So far, we mistakenly assumed that disassociated pools are
automatically cleaned up by Spark. The set of names is initially
allocated in the number of vcores but scheduler pools are only allocated
on demand by parfor workers if in spark execution mode and if there are
large operations in the parfor body.