[SYSTEMML-2240] Fix min/max correctness over inputs w/ NaNs, incl tests
We generally have limited support for NaNs in SystemML because any
sparse operations would be invalid because NaN*0=NaN. However, while sum
and sumSq compute correct results this is not the case for min and max
due to the use of java's basic operators. Instead, we now consistently
use the NaN-aware Math library functions, which so far were only used by
a subset of codegen primitives.

Furthermore, this also includes a cleanup of the builtin function object
that removes the custom execute2 method which is no longer necessary
after we have simplified the exception hierarchy.