[SYSTEMML-2241] Improved selection of ultra-sparse matrix multiply ops
This patch improves the selection of ultra-sparse matrix multiply block
operations (used in all backends) for scenarios where neither of the
inputs qualifies for pure ultra-sparse operations but the output
sparsity is known to satisfy the ultra-sparse sparsity threshold. Since
this ultra-sparse operation is the only matrix multiply operation which
allocates the output in sparse, this patch has the potential to change
the asymptotic behavior.