[SYSTEMML-2245] Support for external UDFs with variable num outputs
This patch adds support for external (i.e., java-based) UDFs with
variable numbers of outputs which is unknown until the UDF has been
executed. In detail, we now allow external UDFs to override the default
assumption of a fixed number of outputs and only verify bound outputs
for UDFs where this is possible.