[SYSTEMML-2252] Improved parfor rewrite update-in-place result indexing
This patch improves the existing update-in-place resulting indexing
rewrite of the parfor optimizer. In detail this includes (1) a fix for
compute the total size of pinned result variables (which incorrectly
double counted the degree of parallelism), and (2) properly compute the
remaining memory estimates without double counting the result indexing
in case this operation is the max memory consumer.

Furthermore, this also includes a cleanup of the parfor cost/memory
estimator to avoid unnecessary tracing and leverage java streams.