[SYSTEMML-2254] Robustness transformdecode for ncol(data) < ncol(meta)
This patch fixes the robustness of transformdecode. So far, we created
the output frame in the number of columns in the given meta data frame.
However, this leads to index-out-of-bound exceptions if the given input
data has fewer columns than the meta data frame due to unnecessary
pass-through encoders that try to access non-existing columns.