[SYSTEMML-1029] Fix correctness sparse-sparse CSR left indexing
This patch fixes incorrect column indexes after sparse-sparse left
indexing into a CSR target matrix (e.g., in case of for/while update in
place variables). In detail, this fast path only applies to aligned
src-tgt indexes as is the case for row-wise indexing but not for general
range-based indexing.