[SYSTEMML-2258] Fix thread contention on acquire shared parfor input
This patch shortens the critical sections of the acquireRead and release
primitives (for buffer pool objects). We now maintain the pinned status
and statistics outside the synchronized block, which improves
unnecessary contention on acquiring shared reads in local parfor loops
with large degree of parallelism. 

As a positive side effect of maintaining the statistics outside the
synchronized block, the -stats output now clearly shows such contention
issues because the waiting time for the lock is counted into the acquire
read time.