[SYSTEMML-1349,2251] Extended parfor optimizer for shared reads
This patch resolves a long standing shortcoming of the parfor optimizer,
of setting the degree of parallelism too low because shared reads were
not handled properly and thus double counted for each worker. Especially
on modern processors with increasing number of cores this became more
important to reduce unnecessary barriers (a too low parfor degree of
parallelism leads to distributing the remaining par to included parfor
loops and operations which require barriers per operation).