[SYSTEMML-2260] New native tsmm matrix mult and its integration
This patch fixes the existing native tsmm implementation to properly use
a BLAS dsyrk instead of BLAS dgemm to performance the computation only
for the upper triangular output matrix. Subsequently, we use the
existing cache-conscious primitive to flip the upper to the lower
triangle. 

Furthermore, this patch also integrates the native tsmm implementation,
which was not used at all so far.