[SYSTEMML-2276] Avoid unnecessary read of empty binary matrices
So far all matrices where read independent of the given number of
non-zeros. For empty, and sparse matrices in general, we allocate a
subset of sparse rows as synchronization pointers. This patch modifies
the binary block readers to early out for known empty matrices and
immediately return a non-allocated matrix block instead. This is
especially helpful for remote parfor loops where empty result variables
are read by all parfor tasks and thus are partially contented on the
shared common thread pool.