[SYSTEMML-2279] Performance spark ctable (1-pass, fused reblock)
This patch significantly improves the performance of the spark ctable
instruction. So far, we constructed double outputs, converted them to
cell, aggregated the cells, determined dimensions, and finally used a
reblock to bring the output into binary block format. For large and
ultra-sparse matrices this caused a redundant pass over the expensive
cell conversion and lots of unnecessary shuffling.

Instead, we now scan the inputs to determine the output dimensions if
necessary, locally aggregate all cells of a partition and directly
output non-zero blocks which feed together with injected empty blocks
(via union) into a fused reblock with global aggregation. In addition,
this also includes numerous smaller improvements to utilize the existing
cluster parallelism and improve memory efficiency and thus reduce the
garbage collection overhead.

On a scenario of sum(table(seq(1,1e9),1+seq(1,1e9)/1000)) with disabled
rewrites and together with the changes from SYSTEMML-2279 through 2282,
this patch improved end-to-end the runtime from 3163s to 455s.