[SYSTEMML-2283] Fix performance issue CSE on DAGs w/ many root nodes
This patch fixes a severe performance issue of the compiler rewrite for
common subexpression elimination (CSE) for the case of DAGs with many
root nodes. The issues showed up on resnet 200 because this script
contains DAGs with 2000+ root nodes. In detail, the issue was due to
incorrect reset-merge of root nodes which led to (predicated on the
graph structure) in the worst-case to processing the entire DAG times
the number of root nodes. This patch fixes the issue and generally
improves performance by using better keys for literal ops (value type
and name) to avoid unnecessary string concatenation. Together, these
changes improved the compilation time of resnet200 from 22min to <10s.