[SYSTEMML-2284] Improved IPA for propagating string args into functions
This patch improves the existing inter-procedural analysis (IPA) for
propagating scalar strings into functions (if it is safe to do so). So
far we only considered long/double/boolean scalars but there are
opportunities for scalar strings (often used as type parameters) because
it allows for predicate folding and branch removal, which are often
prerequisites for additional function inlining.