[MINOR] Performance sparse matrix-vector multiply w/ dense rows
This patch improves the performance of sparse matrix-vector
multiplication for the case where the sparse matrix contains fully dense
rows. The trick is to dispatch dense dot product operations for any
dense row. Since matrix-vector multiplication (dense and sparse) is
memory bandwidth bound, we gain performance by not reading the column
indexes. Even if the matrix fits into caches, this is beneficial by
avoiding unnecessary gather operations.

On a scenario of a 1M x 1K matrix with 30% dense and otherwise empty
rows, this patch improved the MV multiply performance from 140ms to 95s
because we read 8B instead of 12B per non-zero.