[SYSTEMML-2294] Fix robustness constrained parfor optimizer (parameters)
This patch fixes a subtle bug of parameter handling for the CONSTRAINED
parfor optimizer (which optimizes around user-provided forced
parameters). Although all parameters are supposed to be
case-insensitive, when used as "constrained" for example all unspecified
parameters where replaced with defaults which left the constrained
optimizer with no room for decisions.