[SYSTEMML-2296] New sparsity estimator based on matrix histograms
This patch introduces a new sparsity estimator based on matrix
histograms, i.e., a synopsis that stores the number of non-zeros per
row/column for a matrix. This extremely simply yet effective synopsis
allows to exploit structural properties of permutation and selection
matrices. For example, if all rows have <=1 non-zeros the exact number
of non-zeros in the output can be computed as the simple dot product of
lhs columns counts times rhs row counts.