[SYSTEMML-2297] Fix spark left indexing w/ partitioned broadcasts
This patch fixes the distributed spark left indexing operations which
failed due to index-out-of--bound for partitioned broadcasts and
unaligned blocks crossing multiple partitions, while other partitions
are not touched at all. We resolve this by simply pulling the slice over
1k x 1k blocks from the broadcast partition into the partitioned
broadcast because it anyway just works with the getBlock abstraction.