[SYSTEMML-2292,2296] Matrix histograms w/ new exception vectors
This patch improves the general case of sparsity estimation with matrix
histograms where so far we used average case estimates. We now maintain
dedicated exception vectors which enable a hybrid estimator of exact and
approximate fractions. Overall this significantly improves accuracy for
scenarios where a substantial number of rows or columns has <=1
non-zeros (which allow for exact sparsity inference). For an m x n input
matrix, the size of the synopsis remains O(max(n, m)), i.e., in the size
of the largest dimension.