[SYSTEMML-2308,2309] New list data type, incl length, rix, cast ops
This patch introduces a new list data type to group arbitrary data
elements (matrices, frames, scalars, and lists) into a single handle.
Since it's just a meta data handle, we only support the operations
length (number of entries), right indexing (point and range lookups),
casts to matrices, frames, and scalars, and rmvar instructions. 

Note, that this also includes a grammar change and thus regeneration of
the antlr parser.

Furthermore, this includes a minor fix for exist operations to correctly
handle lists and frames.