[SYSTEMML-2325] Fix result correctness bias_add over sparse input
This patch fixes a result correctness issue of the custom bias_add
operation over sparse inputs. This operation used the copy primitive to
take over dense/sparse inputs and subsequently add the bias terms to the
dense output. However, if the input was in sparse format, this led to
switching the output format to sparse such that the bias terms were
added to a discarded dense block and thus lost. We now ensure that the
output is preserved in dense format, which is likely the case for this
additive operation.