[SYSTEMML-2328] Improved codegen row template tmp vector allocation
The row template uses a preallocated ring buffer of temporary vectors
per thread to avoid excessive GC. So far the size of this ring buffer
was determined in a best effort manner by handle unary pipelines and
otherwise counting the number of vector intermediates per operations.
However, for large fused operators this is wasteful as most temporary
vectors are only used for a very localized scope. This patch introduces
a primitive to determine the exact maximum number of live vector
intermediates needed for a fused operator. Overall this improves GC
overheads and cache locality.