[SYSTEMML-2117] Fix missing codegen row vector pow2/mult2 rewrites
This patch adds missing codegen plan rewrites for vector pow(X,2) to
pow2(X) and mult(X, 2) to mult2(X) which were only applied for scalar
operations so far.

Furthermore, this also includes a fix of unnecessary imports and log
instances due to a recent cleanup of the validation phase.