[SYSTEMML-2332] Simplification and performance ALS-predict script
This patch simplifies the ALS-predict script by removing unnecessary,
and even counter-productive script-level "optimizations" which stem from
a time when we did not have sparsity-exploiting fused operators. On a
scenario of a 100Kx100 and 20Kx100 factors and 10K queries, this
modification improved the end-to-end runtime (incl matrix read and
write) from 20.7s to 2.6s.