[SYSTEMML-2350] Fix missing support for lists in as.scalar casts
This patch fixes the missing support for list inputs in as.scalar casts,
which is necessary as a means to index scalars out of unnamed or named
lists because the list indexing itself still returns a list of one
element. Furthermore, this also improves the error handling of the
related runtime instruction.