[MINOR] Fix consistency explain/stats output to stdout (instead of log)
This patch redirects the -explain output to stdout instead of LOG.info
for consistency with the -stats output. On recent HDP versions the log
level is by default set to WARN resulting in the explain output not
being shown. This led to confusion because the stats output was handled
differently.