[SYSTEMML-2353] Fix parfor optimizer block partitioning analysis
This patch fixes an issue of the parfor optimizer that caused a crash
during the block partitioning analysis for special cases of linear
indexing expression with minus and non-existing variables (e.g.,
n-s*i2). Accordingly, we also add a new size propagation test with a
script similar to the failing scenario.