[SYSTEMML-2077] Fix memory estimate second-order eval builtin function
This patch fixes the memory estimate of the second-order eval builtin
function, which is a simple function call and thus has no additional
memory requirements. This is important for the use in parfor loops which
determine the degree of parallelism and other aspects based on the
memory estimates of body operations.