[SYSTEMML-2369] Fix nnz propagation in codegen spark cell ops w/o agg
The recent codegen compiler modification in SYSTEMML-2134 revealed a
hidden issue where spark codegen cell operations without aggregation
wrongly propagated the number of non-zeros of the main input into the
output. This led to incorrect results if the main input is empty but the
cell operation is sparse-unsafe. With this patch we only propagate the
dimensions and blocksizes in this scenario.