[SYSTEMML-2371] New nary min/max operations (language, CP dense/sparse)
This patch introduces new nary min/max operations to avoid unnecessary
nesting of binary operations at script level. As a by-product this also
significantly improves performance if codegen is not enabled.

On a scenario of 10 iterations of min(X, Y, Z, 3, 7) this patch improved
performance compared to the equivalent formulation with binary
operations as follows:
10 x 100K x 1K, dense: 45.1s -> 27.2
10 x 100K x 1K, sparse 0.1: 113.7s -> 8s.

Furthermore, this also includes a refactoring of the existing nary
cbind/rbind operations into a separate package.