[SYSTEMML-2273] Codegen support for nary min/max in cell/row templates
This patch adds operator fusion and code generation support for the new
nary min/max operator in cell and row templates. For both scenarios, we
map the nary operator to binary codegen operations which are anyway
fused on a cell or row level without additional overhead and thus no
need for additional template or runtime primitives.