[SYSTEMML-2383] Fix robustness rowIndexMin/Max for entire rows of NaNs
This patch fixes the unary aggregate operations rowIndexMin and
rowIndexMax for correct handling of rows of all NaNs which so far
resulted in runtime exceptions due to resulting index of -1. We now
define the result as 1 in these cases and also properly reject inputs
with zero columns which would similarly fail.