[MINOR] Performance ternary axpy operations w/ empty inputs
This patch makes a minor performance improvement to the ternary axpy
operations +* and -*. So far, we used a generic and slow fallback
whenever one of the inputs was empty. Instead we now simply use a
shallow copy (i.e., meta data operation) whenever possible. This helps
improve the runtime of paramserv's update function.