[SYSTEMML-2450] Fix parfor optimizer partitioning rewrite, univar stats
With SYSTEMML-2357 we reworked the parfor partitioning rewrites from
statements to hops for correctness and to leverage other compilation
phases. However, this led to a performance regression on perftest univar
stats, 10Mx1K because two instead of one matrices where identified as
partitioning candidates and hence the fused data-partition-execute did
not apply (only valid for single partitioned matrix). 

We now guard this rewrite with a minimal memory constraint based on
local and remote memory budgets (as anyway done during runtime), which
improved the performance of univariate stats 10Mx1K from 452s to 218s
end-to-end.