[SYSTEMML-2452] Fix codegen on parfor recompilation, uni/bivar stats
Dynamic recompilation on parfor entry happens in an in-place manner.
With codegen this is invalid in the presence of changing dimensions
(e.g., stepwise linear regression). Hence, SYSTEMML-2072 introduced
dedicated handling for codegen on in-place recompilation. However, this
led to a performance regression on univariate/bivariate stats with
codegen because the deep copy for codegen happened after clear lops
which corrupted the original DAG without proper recompilation (that
happened on the deep copy). This led to operators being interpreted as
CP and thus to inconsistencies between hops and instructions in case of
large data (i.e., with spark instructions). All this eventually resulted
in a local plan and thus a performance regression. With this patch we
find the right plan again

Furthermore, this also includes a HOTFIX for the previous parfor change
by avoiding the size-based guard in the constrained optimizer as used in
the tests to force data partitioning.