[SYSTEMML-2455] Fix list indexing size propagation on parser-validate
So far only the compiler from hops downward correctly propagated sizes
for list indexing. However, this led to validation issues on parsing for
operations like as.scalar which fails with incorrect input sizes. This
patch does a best effort size propagation for list and otherwise
indicates unknowns which is correct in all cases. Furthermore, this
patch also includes a refactoring of indexed identifier validation to
avoid code duplication.