[MINOR] Improved sparsity estimators bitset, densityMap, matrixHistogram
This patch makes the following minor efficiency and runtime
improvements:

(1) For matrix self-products, we now only construct one instead of two
sketches as the basis for estimation in order to reduce unnecessary
memory and runtime overhead.

(2) The matrix histogram, now leverages information about non-zero lhs
rows and rhs columns to restrict the areas onto which non-zeros are
projected. This technique generally improves accuracy.