[SYSTEMML-2409] Improved matrix histogram (add meta data exploitation)
This patch adds further meta data (the number of rows/cols with 1
non-zero) to the matrix histogram sparsity estimator and exploits this
information during sparsity estimation. Specifically, in mixed mode, the
exact and approximate fractions are known not to collide. Hence, we now
tighten the area the approximate fraction targets which improves the
estimate by more precisely estimating collisions.