[SYSTEMML-2462] Fix correctness/performance in-place sparse binary ops
This patch fixes result correctness issues (or alternatively
index-out-of-bounds exceptions) on in-place sparse binary operations
with left-hand-side inputs in CSR (as produced by special operations or
dense to sparse conversion). This implementation assumed the lhs in MCSR
and hence fails on row merging with CSR and COO. We now ensure that the
in-place target is in MCSR.

Furthermore, this also fixes a couple of performance issues due to
unnecessary copies, unnecessary truncation, and potential gaps after
zero introducing operations.