[SYSTEMML-2463] Function calls w/ named function arguments
This patch generalizes the existing framework of dml-bodied functions by
optionally allowing named function arguments (and thus, permutations of
input arguments). For function calls without name bindings, we simply
use the argument names of the function signature in declaration order.
During compilation and runtime, we then can always assume named bindings
which simplifies the runtime integration.

Furthermore, this patch also generalizes the function inlining in the
parser and during IPA to be robust with regard to permutations of the
input argument list.