[SYSTEMML-2208] Functions with default scalar/matrix/frame parameters
This patch introduces support for default function arguments that are
used whenever the related argument is not provided by a function call.
To this end, this patch modified the parser to allow assignment
expressions in function signatures. Furthermore, we use a simple
compiler approach of adding these default expressions to all function
calls that do not provide certain arguments. Non-provided arguments
without defaults consistently result in compilation (not runtime)
errors.