[SYSTEMML-2291] Performance layered-graph sparsity estimator
This patch significantly improves the build and estimate performance of
the sparsity estimator based on layered graphs (by 1000x and 100x,
respectively). In particular, this includes (1) build graph with
array-based node partitions for O(1) access, (2) reduced node sizes, and
(3) proper memoization on propagating r-vectors bottom-up through the
graph.

On a scenario of estimating the output sparsity of a self matrix product
of a 10K x 10K matrix with sparsity 0.01, this patch improved the
build/estimate times from 798s/120s to 87ms/1.1s.