[SYSTEMML-2296] Performance MNC matrix histogram creation
This improves the creation of matrix histograms in the MNC sparsity
estimator. Specifically this includes (1) fused scans of count vectors
to obtain the summary statistics, (2) a better handling of dense basic
scans, and (3) conditional stores for extension vectors. 

When constructing 4 histograms (1 pair w/o and 1 pair with extension
vectors) for P %*% W, where P:= 58825360 x 2519371 (nnz 58825360) and W
:= 2519371 x 300 (nnz 755810998), this patch improved end-to-end
performance from 14.1s to 10.5s