[SYSTEMML-2492] New broadcast-based spark cumulative aggregates
This patch improves the performance of Spark cumulative aggregates such
as cumsum(X). We now broadcast the offsets in the backwards cascade
whenever possible to avoid shuffling the data input during a repartition
join (unless there exist a partitioner on the main input).

On a 10 node cluster, this patch improved the end-to-end performance of
10 print(sum(cumsum(X))) (without rewrite) over a dense 10M x 2K input
(160GB) from 2011s to 232s.