[SYSTEMML-2509] Fix transform sequences of binning/dummy coding, tests
This patch is a follow-up on fixing the binning support, specifically
for columns that are both binned and dummy coded. For an example
scenario of {recode: [1,2,7], bin: [3,8] dummycode: [3,8]}, we
incorrectly constructed the following composite encoder (which assumed
that all dummy coded columns need to be recoded):

CompositeEncoder(4):
-- EncoderRecode: [1, 2, 3, 7, 8]
-- EncoderPassThrough: [4, 5, 6, 9]
-- EncoderDummycode: [3, 8]
-- EncoderBin: [3, 8]

Now, we fixed that by only adding dummy coded columns that are not
binned to the recode list and brining the basic encoders into the right
sequence (i.e., binning before dummy coding):

CompositeEncoder(4):
-- EncoderRecode: [1, 2, 7]
-- EncoderPassThrough: [4, 5, 6, 9]
-- EncoderBin: [3, 8]
-- EncoderDummycode: [3, 8]

Finally, this patch also includes the necessary tests to ensure such
issues don't occur in the future.