[SYSTEMDS-24] Initial tensor integration (tensor, sum, tostring)
This commit adds the built-in `tensor()` and support for tensors for the
built-ins `sum()` and `toString()`. The implementation is not very clean
and there are a lot of TODOs (which should all be marked by comments).

Note that the CostEstimatorStaticRuntime will probably not be possible
to
accurately work on `tensor()` if we want to support all functionality of
`tensor()` currently implemented.

The `sum()` implementation is also not very in-depth and only very
basic.

`toString()` works nicely, although we might have to think about how we
want to truncate the output if the size is too large. At the moment it
checks for the first two dimensions, but this of course does not really
limit the output size.

Add support for most datatypes on `tensor()`

This additionally fixes a bug in `DenseBlockBool` where the copying of
another DenseBlock did not function correctly.

For all datatypes except `String` it is now supported to fill the tensor
with it, while the tensor keeps the datatype. Also adds print
functionality for all `tensor` containing non double data.
Note that the `String` functionality could not be implemented at the
moment, since we allow data to be passed as list of values in a
`String`, and therefore we can not infer if the user wants to fill the
tensor with the `String` as value or with data in the `String`
concatenated by spaces.
If we pass a `tensor` as an `data`-argument to the constructor, the new
tensor keeps the same datatype.

Closes #17.