[SYSTEMDS-24] Performance tensor aggregations (types, multi-threaded)
This commit improves the sum operation to use multiple threads. This is
as fast as the current matrix implementation for all datatypes except
String, which is slower and it should be discussed how a `sum` on
Strings should even work, and Boolean, where it is a lot faster. If the
matrix is empty though it optimizes it to be a metadata operation (with
non zero count), while the tensor does not yet support this.

Reshape instruction should also be as fast as matrix reshape and makes
use of copy on write.

Note that all instructions are only implemented for CP-mode.

I did some performance tests for a specialized sum implementation for
FP64 (making use of `DenseBlock.valuesAt()`), but got worse performance
for a variety of block sizes (100kx1k, 10kx1k, 1kx1k, 100x1k and 10x1k),
so I will keep the general implementation for now and we can have
another look into it later.

Closes #18.