[SYSTEMDS-24] Modified tensor sum (removed kahan addition)
Replace kahan sum with normal sum for tensor aggregation (uak+
instruction)

This commit additionally contains small bug fix for `DenseBlockString`
and adds a specific `set(DenseBlock)` implementation for
`DenseBlockLString`.

On a two-scoket server with ~90GB/s memory bandwidth this led to a 1.7x
improvement for double and a 3.3x improvement for float. Hence, we
decided to globally remove kahan addition for now. It did not show up in
the past due to servers with lower memory bandwidth and because Kahan
addition has been extended with special handling for NaN and INF.

Closes #27.