[SYSTEMDS-107] Fix correctness IPA check for dimension-preserving
This existing IPA check for dimension-preserving functions was incorrect
in case the input argument was updated inside the function, which by
copy on write led to the reset not taken place. The missing reset then
led to wrong dimensions compiled in the function and based on that
invalid loop bounds and out-of-bounds indexing.