[SYSTEMDS-162] Initial tensor binary instructions (spark, CP)
This should support various binary instructions (+, *, - etc.).
The right hand side tensor (currently only tensor-tensor operations are
allowed) does not have to completely match the left hand side
dimensions, if the dimension is `1` instead. If an dimension is `1` we
reuse this dimension over and over again.
Currently only `repartition join`s are used since size information for
tensors is missing in `hops`, which we would need to decide which
`SPARK` instruction we want to use (`broadcast` or `repartition`). We
could decide between `broadcast` and `repartition` on instruction level,
but this would differentiate from the matrix implementation.

Currently a `Broadcast` for `tensors` can only be triggered by setting
`FORCED_BINARY_METHOD` of `BinaryOp`.

Deciding between `Broadcast` and `Repartition` join for tensors is still
TODO.

Add tensor replication for multiple blocks and minor changes

Closes #31.