[SYSTEMDS-19] Simplify blocksize handling (compiler/runtime), part I
This patch simplifies the handling of block sizes as a precondition for
a deeper integration of tensors. So far, the compiler and runtime
carried row and column blocksizes in order to allow rectangular blocks.
However, in practice only squared blocks were usable (via the external
configuration) and because conditional control flow would require
dedicated handling in all runtime instructions to guarantee alignment
for joins in various spark operations (which they do not have).
Accordingly, we dedicated to restrict ourself to squared blocks which
always guarantee alignment of blocks.