[SYSTEMDS-125] Cleanup and improvements multi-threaded sort
This patch makes some cleanups of the recently added multi-threaded
sort. In detail, this includes a rework of the interaction of
multi-threaded hops with the parfor optimizer and a partially tweaked
implementation of the multi-threaded sort.

On a scenario of 100 1Mx10 (dense) order operations, the previous
multi-threaded sort improved runtime from 32.7s to 22.9s, while with
this patch, we get 16.1s.