[SYSTEMDS-215] List expansion rewrite during dynamic recompilation
This patch introduces a new rewrite that looks for cbind(list) or
rbind(list) and rewrites them to cbind(a,b,c,...n) or rbind(a,b,c,...n),
respectively. While the output dimensions and sparsity of the source
pattern is completely unknown, the latter pattern allows to infer the
size of subsequent operations and thus to avoid unnecessary distributed
spark operations. Since this happens during recompilation, the list can
change during runtime, e.g., as it is the case for cross validation.