[SYSTEMDS-202] Fix rewrite split DAGs after data-dependent ops
This patch fixes an anomaly of incorrect DAGs splits after
data-dependent operations such as removeEmpty, which was caused by
complex dependencies of multiple candidate ops where some candidate was
reachable over other candidates as well as non-candidates. Such
anomalies were already handled, except for the special case where a
candidate directly dependent on the other.

Furthermore, this also includes a minor improvement of the error
handling of variables that are non-existing in the symbol table (ensure
consistent error messages).