[SYSTEMDS-311] Python Language Binding
Add tests, binary operations and federated matrices

We can create matrices from files with `.mtd` or numpy. Basic arithmetic
is implemented and tested, as well as some aggregations.

Federated works, but I have not added tests because we have to start
a worker and we might want to discuss how we distribute the systemds
jar. Currently either a standalone jar or our SystemDS jar + classes
and libs has to placed in the directory `systemds-java` for the
implementation to work, or the java jmlc gateway is manually started
and the method `set_aut_start_jmlc(False)` is called.

I did not yet have a detailed look into packaging and releasing on pip,
but it seems easy with `setuputils`.

Currently on top of todo list are test cases for federated.

For comparisons like `<`, `>=` etc. we also have to talk about either
giving a `DAGNode` back on which `.compute()` has to be called, or
calling `.compute()` for both parts and comparing the results. Further
discussions which actions except `.compute()` actually start the
execution in SystemDS is necessary.

Improve `pre_setup.py` and method method definitions to be similar to
numpy

Calling `pre_setup.py` after building SystemDS
(`mvn package -DskipTests`) will copy `lib` directory and
`systemds-*-SNAPSHOT.jar`, which allows you to run the tests.

Calling `pre_setup.py`, `setup.py sdist` and `post_setup.py`, will
create our package and move it to `target`, which can then be uploaded
to pip (I will look into the specific steps to publish it).

Fix setup for `sdist` (provide jar and so on) and add licenses

Move all files on pre_setup.py to libs

Simplify code and prepare for prep release 0.2

- git ignore java folder for python, and licenses
- Python auto test github actions
- Python Badge

closes #114