[SYSTEMDS-271] Fix compression rewrite (avoid spark context creation)
This patch fixes an issue of the recompile test package
(CSVReadInFunctionTest) where the recently reintroduced compression
rewrite leads to an unnecessary spark context creation. With this fix,
we are know only probing the aggregate cluster memory if we are sure the
matrix does not fit into local memory and hence will be handled as
distributed matrix (which is then subject to compression).